+++ R/utils.R
@@ -589,6 +589,8 @@ circ_mask <- function(d, n, offset = 1) {
   return(dist <= d / 2)
 }
 
+round_dp <- function(x, dp) format(round(x, dp), nsmall = dp)
+
 # equation for following function taken from:
 # https://math.stackexchange.com/questions/2645689/what-is-the-parametric-equation-of-a-rotated-ellipse-given-the-angle-of-rotatio
 
@@ -785,4 +787,44 @@ mean_vec_blocks <- function(x, block_size) {
   }
   
   return(x_out / block_size)
+}
+
+#' Output a table of fit amplitudes and error estimates for a single-voxel fit.
+#' @param fit_res input vector.
+#' @param format_out reduce the accuracy of values to aid table formatting.
+#' @return data.frame of values.
+#' @export 
+sv_res_table <- function(fit_res, format_out = FALSE) {
+  
+  # pretend it's a full fit result if only a table is passed in
+  if (is.data.frame(fit_res)) fit_res <- list(res_tab = fit_res)
+  
+  first_sig        <- names(fit_res$res_tab)[6]
+  first_sig_sd     <- paste0(first_sig, ".sd")
+  first_sig_sd_idx <- which(names(fit_res$res_tab) == first_sig_sd)
+  amp_indx <- 6:(first_sig_sd_idx - 1)
+  amps     <- as.numeric(fit_res$res_tab[1, amp_indx])
+  names    <- colnames(fit_res$res_tab[1, amp_indx])
+  sds      <- as.numeric(fit_res$res_tab[1, amp_indx + 
+                                           amp_indx[length(amp_indx)] - 5])
+  sds_perc <- sds / amps * 100
+  CI95_UB <- amps + sds * 1.96
+  CI95_LB <- amps - sds * 1.96
+  CI95_LB[CI95_LB < 0] <- 0
+  
+  if (format_out) {
+    df_out <- data.frame(amps, sds, sds_perc, CI95_LB, CI95_UB,
+                         row.names = names)
+    df_out <- format(df_out, digits = 3)
+    df_out$sds_perc <- round(sds_perc)
+    df_out$sds_perc[sds_perc > 999] <- "999"
+    df_out$sds_perc <- paste0(df_out$sds_perc, "%")
+    df_out$CI95 <- paste0("[", gsub(" ", "", df_out$CI95_LB), ", ",
+                          gsub(" ", "", df_out$CI95_UB), "]")
+  } else {
+    df_out <- data.frame(amps, sds, sds_perc, CI95_LB, CI95_UB,
+                         row.names = names)
+  }
+  
+  return(df_out)
 }